refactor: importの改善とMakefileのターゲット追加
